components/clp-py-utils/pyproject.toml
[project]
name = "clp-py-utils"
version = "0.7.1-dev"
description = "Utilities for other Python packages in CLP."
authors = [{name = "YScope Inc.", email = "dev@yscope.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.55",
    "mariadb>=1.1.14",
    "mysql-connector-python>=9.4.0",
    "pydantic>=2.12.3",
    "python-Levenshtein>=0.27.1",
    "PyYAML>=6.0.3",
    "result>=0.17.0",
    "sqlalchemy>=2.0.44",
    "StrEnum>=0.4.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
clp = [
    "clp-package-utils",
    "job-orchestration",
]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.4",
]

[tool.mypy]
strict = true

# Additional output
pretty = true
show_error_code_links = true
show_error_context = true
show_error_end = true

[tool.ruff]
extend = "../../ruff.toml"

[tool.uv]
default-groups = ["clp", "dev"]

[tool.uv.sources]
clp-package-utils = { path = "../clp-package-utils" }
job-orchestration = { path = "../job-orchestration" }
